pyproject.toml
[project]
name = "labbench2-pro"
version = "0.1.0"
description = "Scientific LLM evaluation infrastructure"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29",
    "redis[hiredis]>=5.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "anthropic>=0.40",
    "openai>=1.50",
    "google-genai>=1.0",
    "datasets>=3.0",
    "scipy>=1.12",
    "numpy>=1.26",
    "pandas>=2.2",
    "py-irt>=0.6",
    "biopython>=1.83",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]
